Moved a lot into bind.c
