+++ src/bind_utilities.h
@@ -8,17 +8,8 @@
 /** A typedef for bind_lock_t, a simple mutex */
 typedef struct bind_lock_t bind_lock_t;
 
-/** This functions prints Msg and exits the program */
-void bind_err(const char* Msg);
-
-/** This function calls perror then err */
-void bind_p_err(const char* Msg);
-
 /** err if !b */
 void bind_assert(const int b, const char* Msg);
-
-/** This function errors and prints Msg if pnt is null */
-void bind_mem_check(void * pnt, const char* Msg);
 
 /** Malloc, but check to see if failed */
 void * bind_safe_malloc(size_t c, size_t s);
